App/requirements.txt
Flask
google-cloud-storage
scikit-learn
joblib
tensorflow
keras
pandas
numpy==1.16.0
scipy==1.4.1